backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.4.2
pandas==2.1.1
python-multipart==0.0.6
starlette==0.27.0
httpx==0.25.1
colorama==0.4.6 